PROJECTS/TRADING_BOT/requirements.txt
# CRYSTAL: Xi108:W1:A7:S1 | face=R | node=19 | depth=1 | phase=Cardinal
# METRO: Sa
# BRIDGES: Xi108:W1:A7:S2→Xi108:W2:A7:S1→Xi108:W1:A6:S1→Xi108:W1:A8:S1

google-api-python-client>=2.170.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
